Make an OOP based structure as it is getting complicated with multiple monitoring mechanism.
