+++ x-pack/plugins/monitoring/common/constants.ts
@@ -129,6 +129,10 @@ export const INDEX_PATTERN = '.monitoring-*';
 export const INDEX_PATTERN_KIBANA = '.monitoring-kibana-*';
 export const INDEX_PATTERN_LOGSTASH = '*-logstash*';
 export const INDEX_PATTERN_LOGSTASH_MONITORING = '.monitoring-logstash-*';
+export const INDEX_PATTERN_LOGSTASH_STACK_MONITORING_STATE =
+  '.ds-metrics-logstash.stack_monitoring.node-*';
+export const INDEX_PATTERN_LOGSTASH_STACK_MONITORING_STATS =
+  '.ds-metrics-logstash.stack_monitoring.node_stats-*';
 export const INDEX_PATTERN_LOGSTASH_METRICS_NODE = '.ds-metrics-logstash.node-*';
 export const INDEX_PATTERN_LOGSTASH_METRICS_PLUGINS = '.ds-metrics-logstash.plugins-*';
 export const INDEX_PATTERN_BEATS = '.monitoring-beats-*';
