+++ x-pack/plugins/monitoring/server/telemetry_collection/register_monitoring_telemetry_collection.ts
@@ -8,7 +8,7 @@
 import type { IClusterClient } from '@kbn/core/server';
 import type { UsageCollectionSetup } from '@kbn/usage-collection-plugin/server';
 import type { UsageStatsPayload } from '@kbn/telemetry-collection-manager-plugin/server';
-import type { LogstashBaseStats } from './get_logstash_stats';
+import type { LogstashBaseStats } from './logstash_monitoring';
 import type { BeatsBaseStats } from './get_beats_stats';
 import { getAllStats } from './get_all_stats';
 import { getClusterUuids } from './get_cluster_uuids';
